topics routes with protection and admin permissions

+++ app/Http/Controllers/TopicsApiController.php
@@ -70,35 +70,90 @@ class TopicsApiController extends Controller
 
     public function update(Request $request, $id){
 
-        if (Topic::where('id', $id)->exists()) {
-            $topic = Topic::find($id);
-            $topic->title = is_null($request->title) ? $topic->title : $request->title;
-            $topic->save();
+        $isGuest = auth()->guest();
 
-            return response()->json([
-                "message" => "topic updated successfully"
-            ], 200);
+        if(! $isGuest){
+
+            $user_id = auth()->user()->id;
+            $user_role = auth()->user()->role;
+
+
+            if (Topic::where('id', $id)->exists()) {
+
+                $topic = Topic::find($id);
+
+                if($user_id == $topic->user_id || $user_role == 1){
+
+                $topic->title = is_null($request->title) ? $topic->title : $request->title;
+                $topic->user_id = $topic->user_id;
+                $topic->save();
+
+                return response()->json([
+                    "message" => "topic updated successfully"
+                ], 200);}
+                else{
+                    return response()->json([
+                        "message" => "Unauthorized"
+                    ], 401);
+                }
             } else {
             return response()->json([
-                "message" => "topic not found"
+                 "message" => "topic not found"
             ], 404);
+            }
+        }
+        else{
+            return response()->json([
+                "message" => "Unauthorized"
+              ], 401);
         }
     }
 
     public function destroy($id){
 
-        if(Topic::where('id', $id)->exists()) {
-            $topic = Topic::find($id);
-            $topic->delete();
+        $isGuest = auth()->guest();
 
+        if(! $isGuest){
+
+            $user_id = auth()->user()->id;
+            $user_role = auth()->user()->role;
+
+
+            if (Topic::where('id', $id)->exists()) {
+
+                $topic = Topic::find($id);
+
+                if($user_id == $topic->user_id || $user_role == 1){
+
+                    if(Topic::where('id', $id)->exists()) {
+                        $topic = Topic::find($id);
+                        $topic->delete();
+
+                        return response()->json([
+                          "message" => "topic deleted"
+                        ], 202);
+                      } else {
+                        return response()->json([
+                          "message" => "topic not found"
+                        ], 404);
+                      }
+                }
+                else{
+                    return response()->json([
+                        "message" => "Unauthorized"
+                    ], 401);
+                }
+            } else {
             return response()->json([
-              "message" => "topic deleted"
-            ], 202);
-          } else {
-            return response()->json([
-              "message" => "topic not found"
+                 "message" => "topic not found"
             ], 404);
-          }
+            }
+        }
+        else{
+            return response()->json([
+                "message" => "Unauthorized"
+              ], 401);
+        }
     }
 
 }
